requirements.txt
python-vlc
rich
Levenshtein
